# lxdweb/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Session 的认证",
        "title": "LXD Web 管理平台",
        "termsOfService": "https://github.com/xkatld/zjmf-lxd-server",
        "contact": {
            "name": "xkatld",
            "url": "https://github.com/xkatld/zjmf-lxd-server",
            "email": "xkatld@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE"
        },
        "version": "1.0.1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/containers": {
            "get": {
                "description": "从本地数据库获取所有容器缓存信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "获取容器列表",
                "responses": {
                    "200": {
                        "description": "成功返回容器列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/cache": {
            "get": {
                "description": "调用所有活跃节点的lxdapi /api/cache/containers接口快速获取容器信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "从lxdapi缓存获取容器列表",
                "responses": {
                    "200": {
                        "description": "成功返回容器列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/create": {
            "post": {
                "description": "在指定节点创建新的LXD容器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "创建容器",
                "parameters": [
                    {
                        "description": "容器配置参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/{name}": {
            "get": {
                "description": "获取指定容器的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "获取容器详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回容器详情",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点或容器不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/{name}/delete": {
            "post": {
                "description": "删除指定的容器并清理数据库缓存",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "删除容器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/{name}/refresh": {
            "post": {
                "description": "实时获取指定容器的最新信息并更新缓存",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "刷新单个容器信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/{name}/restart": {
            "post": {
                "description": "重启指定的容器",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "重启容器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "重启成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/{name}/start": {
            "post": {
                "description": "启动指定的容器",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "启动容器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "启动成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/containers/{name}/stop": {
            "post": {
                "description": "停止指定的容器",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "容器管理"
                ],
                "summary": "停止容器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "停止成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}